move declaration for DefaultRobotHWSim to header file
